refactor: rename 1.3.3 folders to 1.4.4
