+++ src/wallet-instance-attestation/index.ts
@@ -1,3 +1,3 @@
 export type * from "./api";
 export { WalletInstanceAttestation as V1_0_0 } from "./v1.0.0";
-export { WalletInstanceAttestation as V1_3_3 } from "./v1.3.3";
+export { WalletInstanceAttestation as V1_4_4 } from "./v1.4.4";
