+++ src/wallet-instance-attestation/v1.4.4/utils.ts
