+++ src/wallet-instance/index.ts
@@ -1,3 +1,3 @@
 export type * from "./api";
 export { WalletInstance as V1_0_0 } from "./v1.0.0";
-export { WalletInstance as V1_3_3 } from "./v1.3.3";
+export { WalletInstance as V1_4_4 } from "./v1.4.4";
